Big commit with changes to the eco-system.
Split up the code into model, view, and controller kind of style.
Essentially all business code will be going into the data access objects
and separated from the structural "bean" code.

Changes to the DB with some added fields and adjusted types.

Got the tutor table working, and will be implementing search in the
future.

**This is by far a testing commit, if people don't like the changes we
can revert to a previous commit on the master.